+++ R/compute_posterior_m_nn.R
@@ -17,10 +17,7 @@
 #' @param y_obs_ord A vector of observed outcome values. The vector is ordered
 #' as \code{obs_ord}.
 #' @param kernel_fn The covariance function of the GP.
-#' @param n_neighbor The number of nearest neighbors on one side
-#' (see also \code{expand}).
-#' @param expand Scaling factor to determine the total number of nearest
-#' neighbors. The total is \code{2 * expand * n_neighbor}.
+#' @param n_neighbor The number of nearest neighbors on one side.
 #' @param block_size Number of samples included in a computation block.
 #' Mainly used to balance the speed and memory requirement.
 #' Larger \code{block_size} is faster, but requires more memory.
@@ -42,7 +39,6 @@ compute_posterior_m_nn <- function(hyperparam,
                                    y_obs_ord,
                                    kernel_fn = function(x) exp(-x ^ 2),
                                    n_neighbor = 10,
-                                   expand = 5,
                                    block_size = 1e4) {
 
 
@@ -58,16 +54,16 @@ compute_posterior_m_nn <- function(hyperparam,
 
 
   if(w >= obs_ord[nrow(obs_ord),1]){
-    idx_select <- seq(nrow(obs_ord) - expand * n_neighbor + 1, nrow(obs_ord), 1)
+    idx_select <- seq(nrow(obs_ord) - n_neighbor + 1, nrow(obs_ord), 1)
   }else{
     # which.max returns the index of first TRUE value.
     idx_anchor <- which.max(obs_ord[,1] >= w)
-    idx_start <- max(1, idx_anchor - n_neighbor * expand)
-    idx_end <- min(nrow(obs_ord), idx_anchor + n_neighbor*expand)
+    idx_start <- max(1, idx_anchor - n_neighbor)
+    idx_end <- min(nrow(obs_ord), idx_anchor + n_neighbor)
     if(idx_end == nrow(obs_ord)){
-      idx_select <- seq(idx_end - n_neighbor * 2 * expand + 1, idx_end, 1)
+      idx_select <- seq(idx_end - n_neighbor * 2 + 1, idx_end, 1)
     }else{
-      idx_select <- seq(idx_start, idx_start + n_neighbor * 2 * expand - 1, 1)
+      idx_select <- seq(idx_start, idx_start + n_neighbor * 2 - 1, 1)
     }
   }