+++ R/estimate_mean_sd_nn.R
@@ -16,10 +16,7 @@
 #'   (e_gps_pred).
 #'   - Column 3: Standard deviation of  e_gps (e_gps_std)
 #' @param kernel_fn The covariance function of the GP.
-#' @param n_neighbor The number of nearest neighbors on one side
-#' (see also \code{expand}).
-#' @param expand Scaling factor to determine the total number of nearest
-#' neighbors. The total is \code{2 * expand * n_neighbour}.
+#' @param n_neighbor The number of nearest neighbors on one side.
 #' @param block_size The number of samples included in a computation block.
 #' Mainly used to balance the speed and memory requirement. Larger
 #' \code{block_size} is faster, but requires more memory.
@@ -39,7 +36,6 @@ estimate_mean_sd_nn <- function(hyperparam,
                                 GPS_m,
                                 kernel_fn = function(x) exp(-x ^ 2),
                                 n_neighbor = 50,
-                                expand = 2,
                                 block_size = 2e3,
                                 nthread = 1){
 
@@ -72,7 +68,7 @@ estimate_mean_sd_nn <- function(hyperparam,
                           varlist = c("w", "GPS_m", "hyperparam",
                                       "coord_obs_ord", "y_use_ord",
                                       "sigma2", "kernel_fn",
-                                      "n_neighbor", "expand", "block_size",
+                                      "n_neighbor", "block_size",
                                       "compute_posterior_m_nn",
                                       "compute_posterior_sd_nn",
                                       "compute_inverse", "calc_cross"),
@@ -93,7 +89,6 @@ estimate_mean_sd_nn <- function(hyperparam,
                                    sigma2 = sigma2,
                                    kernel_fn = kernel_fn,
                                    n_neighbor = n_neighbor,
-                                   expand = expand,
                                    block_size = block_size)
     val
   })